+++ lib/cli.js
@@ -1,5 +1,23 @@
 #!/usr/bin/env node
 
-import { run } from './index';
+import whiten from './index';
 
-run(process.argv);
+
+function showUsage() {
+  console.log([
+    'Maven Whiten',
+    '------------',
+    '',
+    'mvn-whiten <groupId>:<artifactId>:<version>',
+  ].join('\n'));
+}
+let packages = process.argv.slice(2);
+if (args.length === 0 || args.indexOf('--help') !== -1 || args.indexOf('-h') !== -1) {
+  return showUsage();
+}
+
+let { tar, cd } = whiten( packages );
+let filePath = packages.join('-').replace(/:/g, '_') + '.tar';
+tar.pipe(fs.createWriteStream(filePath)).on('end', () => {
+  cb();
+});
